refactor : 쉐어 작성 페이지 컴포넌트 분리  #102
